7. Проект 1/Subscriptions_Dc_Pstg_TG_Clr_Flw/requirements.txt
Django==5.0.1
djangorestframework==3.15.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
flower==2.0.1
python-decouple==3.8
django-environ==0.11.2
django-celery-results==2.6.0
django-celery-beat==2.6.0
pyTelegramBotAPI==4.16.1
sqlalchemy==2.0.25
psycopg2==2.9.9